fix 32-bit build

+++ src/rpc/system.rs
@@ -894,8 +894,8 @@ impl NodeStatus {
 		use nix::sys::statvfs::statvfs;
 		let mount_avail = |path: &Path| match statvfs(path) {
 			Ok(x) => {
-				let avail = x.blocks_available() * x.fragment_size();
-				let total = x.blocks() * x.fragment_size();
+				let avail = x.blocks_available() * x.fragment_size() as u64;
+				let total = x.blocks() * x.fragment_size() as u64;
 				Some((avail, total))
 			}
 			Err(_) => None,
